Added the first heuristic

+++ src/graph/Graph.java
@@ -12,6 +12,7 @@ public class Graph {
 	}
 	
 	public void generateGraph(int nbVertices, double redVertexProb, double blueEdgeProb) {
+		vertices.clear();
 		for(int i = 0; i < nbVertices; i++) {
 			addVertex(redVertexProb, blueEdgeProb);
 		}
@@ -30,7 +31,9 @@ public class Graph {
 		new Edge(getColorFromProb(1-blueEdgeProb), v, newVertex);
 	}
 	
-	public void removeVertex(Vertex v) {
+	public void removeVertex(Vertex v) throws Exception {
+		if(v.getColor() == Color.BLUE)
+			throw new Exception("TRYING TO DELETE BLUE VERTEX");
 		vertices.remove(v);
 		v.delete();
 	}
@@ -50,6 +53,9 @@ public class Graph {
 				System.out.println(tripleTab+"\t---"+(e.getColor() == Color.RED ? "R" : "B")+"--->"+e.getDst().color);
 			}
 		}
-
+	}
+	
+	public List<Vertex> getVertices() {
+		return vertices;
 	}
 }
